refactor(shared): separate `schemas/admin.ts` into several files
