+++ shared/src/index.ts
@@ -1,2 +1,6 @@
-export * from "./schemas/admin.js";
+export * from "./schemas/admin/adminCreatePostResult.js";
+export * from "./schemas/admin/adminPostDetails.js";
+export * from "./schemas/admin/adminPosts.js";
+export * from "./schemas/admin/adminSubmitPost.js";
+export * from "./schemas/admin/adminUsers.js";
 export * from "./schemas/posts.js";
