+++ shared/src/schemas/admin/adminSubmitPost.ts
@@ -1,0 +1,15 @@
+import * as v from "valibot";
+import { nanoidString, positiveInteger, trimmedString } from "../types.js";
+
+const adminSubmitPostSchema = v.object({
+  authorId: nanoidString,
+  timestamp: positiveInteger,
+  caption: trimmedString,
+  reactions: positiveInteger,
+});
+
+export type AdminSubmitPostInput = v.InferInput<typeof adminSubmitPostSchema>;
+
+export function parseAdminSubmitPost(input: unknown) {
+  return v.parse(adminSubmitPostSchema, input);
+}
