+++ shared/src/schemas/admin/adminUsers.ts
@@ -1,0 +1,16 @@
+import * as v from "valibot";
+import { integerToBoolean, nanoidString, trimmedString } from "../types.js";
+
+const adminUsersSchema = v.array(
+  v.object({
+    id: nanoidString,
+    name: trimmedString,
+    hasAvatar: integerToBoolean,
+  }),
+);
+
+export type AdminUsersInput = v.InferInput<typeof adminUsersSchema>;
+
+export function parseAdminUsers(input: unknown) {
+  return v.parse(adminUsersSchema, input);
+}
